+++ src/elements/LazyImage/index.js
@@ -1,47 +1,78 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import InlineSVG from 'svg-inline-react'
+import 'lazysizes'
 
+import { Dimension, Dimensions } from './Dimension'
 import Spinner from './spinner.svg'
 import './style.scss'
 
-const Dimensions = {
-  '16x9': (9 / 16) * 100
-}
-export default class LazyImage extends React.Component {
+const imageUrlPrefix = 'https://res.cloudinary.com/spencerkline/image/upload'
+
+export class LazyImage extends React.Component {
   static defaultProps = {
     alt: '',
     className: '',
-    dimension: undefined,
+    dimension: Dimensions.normal,
     placeholder: <InlineSVG src={Spinner} raw={Element.prototype.hasOwnProperty('remove')} />
   }
   static propTypes = {
     src: PropTypes.string.isRequired,
     alt: PropTypes.string,
     className: PropTypes.string,
-    dimension: PropTypes.oneOf(Object.keys(Dimensions)),
+    dimension: PropTypes.instanceOf(Dimension),
     placeholder: PropTypes.node,
-    rounded: PropTypes.bool
+    portraitRotation: PropTypes.string
   }
   constructor (props) {
     super(props)
     this.state = {
-      isLoaded: false
+      isLoading: true
     }
-    let image = new Image()
-    image.onload = () => {
-      this.setState({ isLoaded: true })
-    }
-    image.src = this.props.src
   }
+
   render () {
-    const {src, alt, className, dimension, placeholder, rounded} = this.props
+    const {className, dimension, placeholder, portraitRotation} = this.props
     return (
-      <div className={className + ' lazyImage'} style={dimension ? {paddingBottom: Dimensions[dimension] + '%'} : {}}>
-        {this.state.isLoaded
-          ? <img src={src} alt={alt} style={dimension ? {height: '100%'} : {}} className={rounded ? 'br2' : ''} />
-          : placeholder}
+      <div className={className + ' lazy-image'} style={portraitRotation ? {} : { paddingBottom: dimension.ratioPercentage + '%' }}>
+        { this._responsiveImage() }
+        { this.state.isLoading ? placeholder : null }
       </div>
     )
   }
+
+  /**
+   * Generates an HTML5 picture tag with multiple image sources responsive to browser and screen size (saving bandwidth)
+   *
+   * @return <picture> element containing x sources (x = number image types) and 1 fallback img. Each source has a
+   * number of srcSet options equal to number of possible image widths.
+   */
+  _responsiveImage () {
+    const {alt, className, dimension, portraitRotation, src} = this.props
+    const imageTypes = [
+      { extension: 'webp', type: 'image/webp' },
+      { extension: 'wdp', type: 'image/vnd.ms-photo' },
+      { extension: 'jpg', type: 'image/jpg' }
+    ]
+    const prefix = imageUrlPrefix + (portraitRotation ? '/a_' + portraitRotation : '')
+    return (
+      <picture>
+        { imageTypes.map(({type, extension}) => {
+          // For each image type, add a source tag
+          return (
+            <source key={extension} type={type} srcSet={dimension.widths.map((width) => {
+              return `${prefix}/c_scale,w_${width}/${src}.${extension} ${width}w`
+            })} />
+          )
+        }) }
+        {/* Fallback image for older browsers */}
+        <img data-src={`${prefix}/${src}.jpg`} alt={alt} className={className + ' lazyload'} onLoad={() => {
+          if (this.state.isLoading) {
+            this.setState({ isLoading: false })
+          }
+        }} />
+      </picture>
+    )
+  }
 }
+export { Dimensions } from './Dimension'